requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
aiohttp
alembic
pydantic
fastapi-admin